Move hdinsight/2018-06-01-preview out of preview since it should be stable
