+++ CHANGELOG.md
@@ -39,6 +39,8 @@
 
 - package `github.com/Azure/azure-sdk-for-go/services/cosmos-db/mgmt/2020-04-01/documentdb` has been renamed to `github.com/Azure/azure-sdk-for-go/services/preview/cosmos-db/mgmt/2020-04-01-preview/documentdb`
 
+- package `github.com/Azure/azure-sdk-for-go/services/preview/hdinsight/mgmt/2018-06-01-preview/hdinsight` has been renamed to `github.com/Azure/azure-sdk-for-go/services/hdinsight/mgmt/2018-06-01/hdinsight`
+
 Generated from https://github.com/Azure/azure-rest-api-specs/tree/93106ff722d481c30a9a4ae089564635c10e9401
 
 ## `v46.4.0`
