+++ services/hdinsight/mgmt/2018-06-01/hdinsight/applications.go
