+++ services/hdinsight/mgmt/2018-06-01/hdinsight/hdinsightapi/interfaces.go
@@ -19,7 +19,7 @@ package hdinsightapi
 
 import (
 	"context"
-	"github.com/Azure/azure-sdk-for-go/services/preview/hdinsight/mgmt/2018-06-01-preview/hdinsight"
+	"github.com/Azure/azure-sdk-for-go/services/hdinsight/mgmt/2018-06-01/hdinsight"
 	"github.com/Azure/go-autorest/autorest"
 )
